[BE] fix : entity 수정

See merge request s07-webmobile1-sub2/S07P12D105!57

+++ BE/sellerB/src/main/java/backend/sellerB/controller/NoticeController.java
@@ -53,7 +53,7 @@ public class NoticeController {
     }
 
     @GetMapping("/{seq}")
-    public ResponseEntity<NoticeDto> getNoticeDetail(@PathVariable Integer seq) {
+    public ResponseEntity<NoticeDto> getNoticeDetail(@PathVariable Long seq) {
         return ResponseEntity.ok(noticeService.getNoticeDetail(seq));
     }
 
@@ -64,7 +64,7 @@ public class NoticeController {
 
 
     @PutMapping("/{seq}")
-    public ResponseEntity<NoticeDto> putNoticeDetail(@Valid @RequestBody NoticeDto noticeDto, @PathVariable Integer seq) {
+    public ResponseEntity<NoticeDto> putNoticeDetail(@Valid @RequestBody NoticeDto noticeDto, @PathVariable Long seq) {
         return ResponseEntity.ok(noticeService.update(seq, noticeDto));
     }
 
@@ -77,7 +77,7 @@ public class NoticeController {
 
     //soft delete
     @DeleteMapping("/{seq}")
-    public ResponseEntity<NoticeDto> deleteNotice(@PathVariable Integer seq) {
+    public ResponseEntity<NoticeDto> deleteNotice(@PathVariable Long seq) {
         // Access the DB and delete the order
         return ResponseEntity.ok(noticeService.deleteNotice(seq));
     }
